requirements.txt
cmake
dlib==19.18.0
face-recognition
numpy
opencv-python
PyQt-builder
mysql.connector